+++ whitelist-authenticator/src/main/java/de/intension/authentication/WhitelistAuthenticatorFactory.java
@@ -1,20 +1,21 @@
 package de.intension.authentication;
 
-import java.util.List;
-
 import org.keycloak.Config;
 import org.keycloak.authentication.Authenticator;
 import org.keycloak.authentication.AuthenticatorFactory;
+import org.keycloak.constants.AdapterConstants;
 import org.keycloak.models.AuthenticationExecutionModel;
 import org.keycloak.models.KeycloakSession;
 import org.keycloak.models.KeycloakSessionFactory;
 import org.keycloak.provider.ProviderConfigProperty;
 
+import java.util.List;
+
 /**
  * Factory to create custom {@link WhitelistAuthenticator}.
  */
 public class WhitelistAuthenticatorFactory
-    implements AuthenticatorFactory
+    implements AuthenticatorFactory, AdapterConstants
 {
 
     public static final String           PROVIDER_ID            = "whitelist-authenticator";
@@ -73,7 +74,11 @@ public class WhitelistAuthenticatorFactory
     public List<ProviderConfigProperty> getConfigProperties()
     {
         return List.of(
-                       new ProviderConfigProperty(LIST_OF_ALLOWED_IDP, "Whitelist of IdPs", "Configuration of allowed IdPs for specific clients.",
+                       new ProviderConfigProperty(IdpHintParamName.IDP_HINT_PARAM_NAME, "IdP hint parameter name",
+                               "Name of the URL query parameter which contains the allowed IdP.",
+                               ProviderConfigProperty.STRING_TYPE, KC_IDP_HINT),
+                       new ProviderConfigProperty(LIST_OF_ALLOWED_IDP, "Whitelist of IdPs",
+                               "Configuration of allowed IdPs for specific clients.",
                                ProviderConfigProperty.TEXT_TYPE, null));
     }